feat(pipeline): Add a step to the pipeline (build -> deploy) which notifies Slack when a manual review is required
This commit introduces a series of steps in the pipeline (test -> build
-> deploy) GitHub action which gathers some basic information about the
(previously-merged) PR, and sends a HTTP POST request to a known Slack
URL.

Deployment gating / manual approval of deployments used to take place in
CircleCI but now happens in GitHub actions.

We want to create a publically accessible, and push-based alert system
to let developers know that their deployent is ready for testing.

We want to do this to reduce the time spent between a PR being merged to
`main` and a deployment going out.

Building a custom Slack Workflow URL appears to be the best way to
achieve this.