APG-1213(fix): Remove the lines from the TestDataCleaner which remove the regions, pdus, and office
**WHAT**

This fixes the issue where, when tests are run in sequence to each
other, the reference data for regions, pdus, and offices are being
clearned out.

Because this data is not user created (i.e. it will not be writeable
except by our migration script, at least in its current iteration), i
have simply elected to remove the lines from the Clean Up Script which
remove it - because the data will be created (by database migrations),
in the exact state that it will be used.

This approach may well need to change in the future, if any of this data
ever becomes more dynamically generated.