APG-1222: Create nDeliusIntegrationApi::getRequirementManagerDetails
**WHAT**

This is a WIP-commit, it does not contain a full feature.

This commit introduces a method to integrate with the GET
/case/{crn}/requirement/{requirementId} URL on the nDelius API to help
return the ProbationDeliveryUnit (PDU) associated with a Person on
Probation's Probation Practitioner.

This commit builds out the basic integration, and creates the relevant
wiremock response, as well as an integration test.

It also introduces a set of data classes, with TODO comments (signed by
the author) to highlight the limits of WIP.

There are

**WHY**

This endpoint will eventually be used to provide a suggested list of Offices
(obtained _via_ the PDU) to the user of the "Set the Delivery Location
Preferences" for a Person on Probation in relation to a specific Referral.

Linking a Person on Probation to a geographic region is a known
difficult problem in the community setting.  By keeping our commits small, we can learn as we go.