APG 1213/create region pdu office reference data (#128)
* APG-1213: Create the region, pdu, and office data tables

**WHAT**

This PR creates three new tables, their entities, and their repositories, to contain the reference data for Delivery Location Preferences

It includes documentation explaining these data models (``/docs/schema-docs/2025-08-26-geographic-reference-data.md`)

* APG-1213(migrations): Add loading data from regions, pdus, offices.

* APG-1213(chore): Klint

* APG-1213(fix): Remove duplicate tail line from the load_office_data.sql file

* APG-1213(fix): Replace 'True' and 'False' with 'true' and 'false' in the load_office_data SQL script

* APG-1213(migrations): Rename various id-related fields in region, pdu, office

* APG-1213(repositories): Remove manual queries for (unused) repo methods on pdu, office, region

* APG-1213(tests) Add repository-levle tests for Region, Office, PDU

* APG-1213(fix): Remove the lines from the TestDataCleaner which remove
the regions, pdus, and office

**WHAT**

This fixes the issue where, when tests are run in sequence to each
other, the reference data for regions, pdus, and offices are being
clearned out.

Because this data is not user created (i.e. it will not be writeable
except by our migration script, at least in its current iteration), i
have simply elected to remove the lines from the Clean Up Script which
remove it - because the data will be created (by database migrations),
in the exact state that it will be used.

This approach may well need to change in the future, if any of this data
ever becomes more dynamically generated.