APG-1170: Create the referral::sourced_from field
**WHAT**

This commit introduces the sourced_from field onto the ReferralEntity,
and into the referral table, via a flyway migration.

It updates the data class (`FindAndReferReferralDetails`) by simply
copy-pasting from the source.

**WHY**

In order to source data about the Delivery Locations for a Referral, we
must determine if it was created from a Requirement or from a Licence.
This has implications about how exactly to link a Referral to a
Probation Delivery Unit, via the following pathway:

```
Referral -> Requirement|Licence -> Probation Practitioner -> PDU ->
Offices
```

Although the F&R presently sends the ID of the Licence or Requirement
(which we call the `eventId`, which may have proven a bit of a misnoma),
we didn't actually store the sourced_from field.  I have created this
field, on the entity and in the database, so that we might store it.

In a follow-up commit I will create a SQL migration which sets the
sourced_from field on all existing Referrals, through some manual
exports from the F&R database, as this is likely the most pragmatic,
easy way forward.