APG-1222: Create the ReferralService::attemptToFindManagerForReferral method
**WHAT**

This commit introduces a new method into the ReferralService, which first attempts to find the Manager for a Requirement, and then attempts to find it for the Licence, associated with a Referral.

**WHY**

In order to deliver a set of Delivery Locations to a user, associated with a Referral, we need to search based on the Case, associated with the Person on Probation and the upstream "sourced from" event, i.e. a Licence Condition or a Requirement.

Although we _may_ have some knowledge about if our event was sourced from a Requirement or Referral, this method assumes a naive system, that has to check both.