chore(NDeliusApiIntegration): Update the shape of the Requirement or Licence Condition fetching
**WHAT**

This commit introduces a handful of data classes to represent the data that comes back from either of the following NDelius Integration URLs:

- `GET /case/$crn/requirement/$requirement_id`
- `GET /case/$crn/licence-conditions/$licence_condition_id`

It adds the `probationDeliveryUnits` as a root-level field (a sibling to the `manager`).

**WHY**

This was requested from the Probation Integration team so that we could (possibly) show the list of Offices / Delivery Locations outside the "primary" PDU of the Manager.

We did this because there is an (optional) step in the UI which would allow a user to select one of these locations.

**TODO**

Updating tests and mocks, will come in the next commit.