fix(DeliveryLocationPreferences): Add the '.code' field to PDU
**WHAT**

This commit adds the missing `.code` field (String) to the PDU being
returned for the DeliveryLocationPreferencesFormData

**WHY**

This code is necessary to save the PDU into the database when
creating/updating the DeliveryLocationPreferences in the Repository.

Without it, there would be un-necessary re-fetching of data and matching
based on `name` or `description` which feels unnecessarily complex and
risky.