APG-1339: Update the Domain Model around Referral and Referral Status History
**WHAT**

This commit introduces a core change to the modelling of a Referral's Status

It removes the previous mapping table implementation, and now just links from a Referral to a ReferralStatusHistory, via a `referral_id`

Additionally, a Status is no longer an optional field on the Referral, and mandating up this relationship required a handful of changes.

**WHY**

The previous implementation was made before we had our current understanding of Statuses.

The previous implementation introduced unnecessary complexity (an unnecessary database table), we so can achieve the same thing in less code.