Create an endpoint to update the ReferralStatusHistory (#168)
* APG-1339: Update the Domain Model around Referral and Referral Status History

**WHAT**

This commit introduces a core change to the modelling of a Referral's Status

It removes the previous mapping table implementation, and now just links from a Referral to a ReferralStatusHistory, via a `referral_id`

Additionally, a Status is no longer an optional field on the Referral, and mandating up this relationship required a handful of changes.

**WHY**

The previous implementation was made before we had our current understanding of Statuses.

The previous implementation introduced unnecessary complexity (an unnecessary database table), we so can achieve the same thing in less code.

migrations: remove duplicates, and fix the order of all the migrations

APG-1339: Fix case sensitive Awaiting Assessment -> Awaiting assessment

APG-1339: Rename V27 migration to V28, for PR#165

APG-1339: fix the last case sensitive Assessment -> assessment

* APG-1339: Create the POST /referral/:id/status-history endpoint

**WHAT**

This commit creates a new endpoint at `POST /referral/:id/status-history` which allows a user to update the Referral Status of a Referral by creating a new entry in the Referral Status History table

**WHY**

This endpoint will eventually be used by the UI in order to update the
Referal Status.

* fix(ReferralStatusDescriptionRespositoryIntegrationTest): Remove unused argument from constructor

* APG-1339: Update the CreateReferralStatusHistory endpoint to take additionalDetails; add controller-level test

* remove the ReferralStatusDescriptionRepositoryIntegrationTest

* ktlint

---------

Co-authored-by: pfcairns <phil.feroldi-cairns@digital.justice.gov.uk>