feat(ReferralReportingLocation): Add `region_name` findByReferralId (#211)
**WHAT**

This commit introduces the `region_name` (sql column) or `regionName` (entity) field to the Reporting Location entity.

**WHY**

In order to filter Referrals to a logged-in user's Region, we need to be able to associate a Referral to a Region, which the ReportingLocation did not previously do.

**REFACTORING**

I have added in additional stubbing and assertions to the
`ReferralServiceIntegrationTest.kt` file, which now assert on the
creation of the ReportingLocation.  This behaviour was previously
un-tested.