feat(groups): Combine Cohort and LDC in Group Filter
**WHAT**

This PR combines the concepts of a _Cohort_ (e.g. GENERAL vs. SEXUAL)
and `hasLdc` (representing learning difficulties and challenges) into
one set of values from the `GET /bff/group/{groupId}/{selectedTab}` endpoint.

These values have been added to the
`.allocationAndWiatlistData.filters.cohort` field.

These values can now be received, and processed, on GET so that the
relevant Referrals in the Waitlist are displayed.

**WHY**

Although our system separates LDC and Cohort as two separate fields,
User Research in Custody (and therefore in Community) indicates that
some users consider "Cohort" to actually be a combination of these two
factors.

Therefore, in several dropdown/select filters, we are required to show
the following options:

- `General Offence`
- `General Offence - LDC`
- `Sexual Offence`
- `Sexual Offence - LDC`

**HOW**

We have delegated this decision to the backend in Community in our
efforts to keep the UI an incredibly unopinionated "display only" layer,
which displays the information presented to it.

The newly created `ProgrammeGroupCohort` enum class generates the
above-given labels, and is able to convert them from (and to) the 2x2
matrix of Cohort and LDC values.  This happens early on in the request's
lifecycle (i.e. as soon as the request is received by the Controller).
This allows for minimal changed code elsewhere.

**TESTING**

This commit also makes some changes to the testing of the controller:

1. It removes duplicate assertions, likely present from copy-paste
2. In order to de-couple the HTTP response from the Kotlin/JVM code, it
   makes assertions on the JSON response, and checks for the literal
   presence of Strings there.  This reduces the risk of code testing
   itself.  From the UI's perspective, the important observable
   behaviour is that a series of strings are received and communicated.