feat(GroupDetails): Add fields & sort by sentenceEndDate
**WHAT**

This commit introduces two changes:

1. Add two fields (`referralId` - a UUID; and `sourcedFrom` - a
  human-readable string of `"Requirement"` or `"Licence Condition"`) to
  the Referral coming through in the CaselistItem
2. Sort CaselistItems by `sentenceEndDate`, ascending - showing
   Referrals with an earliest Sentence End Date first in the Allocated
   and Waitlist table

**WHY**

In order to allow the user to navigate to the Referral Details screen we
need the Referral

In order to display the Waitlist & Allocated tables correctly as per the
Figma, then we need to have the `sourced_from` field being fed through.
As-per our own BFF guidelines, we're doing the string-level
transformation here in the API, not in the UI, so that we can keep the
UI as business uninformed as possible

In-table sorting is on the current page only (a limitation of MOJ and
central gov design Table components) and therefore it's important that
we provide a "useful" sort order to the data.  Speaking to Kevin (BA) it
became apparent that users will often want to prioritise People on
Probation who have a sooner Sentence End Date into a group, so that
people complete their requirements before the end of their Sentence

**TESTING**

- The commit has added a new test to the ProgrammeGroupController which asserts on
  the JSON body of the response for the waitlist, and asserts on the
  literal string returned that represents a date.  This allows for
  explicit expectation setting for date
  transformation/getReferralStatusDescriptionFunction
- The commit adds to an existing test to assert that both the
  `referralId` and the `sourcedFrom` fields of a Referral _are_ being
  passed along to the user