Docs: Accredited Programmes Template
**WHAT**

This commit introduces a technical and semi-architectural document on
the idea of an `AccreditedProgrammeTempalte` as a new entity in our
system, to allow core functionality around creating Sessions for
participants in a Group who complete an Accredited Programme.

**WHY**

We have an upcoming chunk of work around how, exactly, we are to create
the relevant Sessions, and record Attendance, for Members of a Group
completing an Accredited Programme.

There is a lot of complexity around this topic, and although the team
would usually prefer a design and complexity to evolve naturally out of
user needs and functionality, in this case some up-front thinking may
help us form the shape of a solution.

It is my (Wilson's) hope that by doing this thinking up front, and by
presenting the ideas to the Dev team (and then feeding back and
re-shaping them) that we will be able to more smoothly deliver the
upcoming 4-6 weeks' worth of work, without the need to reinvent
solutions in multiple places.