APG-1503: Add is_continuing field to the ReferralStatusTransitionEntity
**WHAT**

This commit introduces the `is_continuing` or `isContinuing` field to
the ReferralStatusTransition(Entity)

**WHY**

This is necessary to make some core decisions in the "Remove a PoP from
a Group" flow.  We need to be able to distinguish between if the Status
Change is "good" (insofar as the Person on Probation is progressing
towards a completed Referral).

Specifically when doing this from the "Status History" tab of the
Referral Details screen.

There is a difference between somebody moving to "Awaiting Allocation"
from "Awaiting Assessment" and moving _back_ to "Awaiting Allocation"
from "Scheduled".  We need some way of having our system distingusih
between these two things.

This is laid out in the [Excalidraw](https://excalidraw.com/#room=88ce0e447fe1cfefdffd,b_vXhLp6O_nLtesi81cAGQ) associated with the JIRA ticket.

**HOW**

Created the field, and default it to `false`.  Set it to `true` on
Status Changes which indicate a forward-movement.

Added some integration-level tests to make sure this is plugged in.