APG-1514: Change status on Allocate; remove membership on Status Change
**WHAT**

This commit introduces two feature-level changes:

1. Implemented removal of a Referral from a ProgrammeGroup when
   transitioning Status in a "non-continuing" journey.
2. Add a StatusHistory entry when a Referral is Allocated to a
   ProgrammeGroup (into "Scheduled")

**WHY**

Because the user will always need to do these things together (e.g.
Allocation to a Group should also _always_ update the Status; and Removal
from a Group should _always_ take the users back to "Awaiting Allocation" Status) -
this code keeps the state of our Referrals in sync as the user does
the actions they are focusing on.

By having the code do this in automated way, we reduce User effort, and
increase data accuracy.

**HOW**

We had to update the
`ProgrammeGroupMembershipService::allocateReferralToGroup` method to
create a ReferralStatusHistoryEntity

Update the `ReferralService::updateStatus` method to conditionally
remove a ProgrammeGroupMembership depending on the from & to status
combination.

**WHAT NEXT**

Surface the "Remove from Group" action/intention via a new Endpoint

This includes the GET /bff endpoint (APG-1504) and POST (APG-1507) to
allow the user to remove a Person on Probation from a Group.