APG-1422: Record AdditionalDetails on GroupAllocation Status Change
**WHAT**

This PR feeds through the "Additional Details" (free form text from)
the User, on Group Allocation of a Referral.

**WHY**

When somebody is allocated to a Group we should change that
Referral's Status to "Scheduled" which also requires the ability for a
User to provide "Additional Details" to record any relevant context
which they may wish to pass on to their team.

**HOW**

A previous PR introduced the ability for the Service method to record
the Additional Details which Allocating a Referral to a Group.  This
PR's scope was thus limited to feeding through the additional details
from the API.

Additionally, the API generates the success message to display to the
User on the Group Details screen, which means that we can take that
decision away from the API.

API types and schemas have also been added and documented.