APG-1504: Create /bff/remove-from-group/:referral_id endpoint (#288)
**WHAT**

This commit introduces a new endpoint: `GET
/bff/remove-from-group/:referral_id`, which returns data of an identical
shape to the existing `GET /bff/status-tranisitions/:referral_id` but
int the context of the "Remove Referral from Group" form.

**WHY**

We did not want to use the same endpoint for both use-cases, because a
change in one will not necessarily require a change in the other.
Therefore this change introduces the minimal change necessary to
duplicate existing functionality, but separate the data return types,
which will allow co-evolution of both contexts without affecting the
other.