APG-1507: Create the endpoint to remove a Referral from a Group (#287)
* refactor: Move existing ProgrammeGroupMemberhsipServiceIntegrationTest units into a nested block

* APG-1507: Create ::removeFromGroup method

This commit introduces the `ProgrammeGroupMembershipService::removeFromGroup` method.

That method does some sense-checking on data (the Referral, the Group, the current Membership), and then does two things: soft-deletes the membership & updates the Referral Status to a Status provided

In a (probably follow-up) step to this work, we need to allow the User to Remove a Person on Probation from a Group.  They will do this using a special screen on the UI, through which they will be presented with the list of possible Statuses they wish to select for the Referral.

In the interest of continuous delivery and incremental features, this PR introduces a functional change separate from the API plumbing.  This allows the focus to be on the functionality and on the testing.

The central logic is all stored in the `ProgrammeGroupMembershipService::removeFromGroup` method.  This makes a relatively procedural checks through data presence and throws any relevant errors.

It then removes any `ProgrammeGroupMembeership` entities (soft deletes them, for our records) and then creates a new record in the `StatusHistory` log, with the required information.

Plugging this Service method into an endpoint (likely `POST /referra/{id}/remove-from-group`), with the relevant Controller / endpoint tests is the clear next step.  Then plugging this into the UI screens which are being spun up.

As part of this commit I have made several tidy-up changes:

- Extracted `::getCurrentlyAllocatedGroup` from the `ReferralService`, and into the `ProgrammeGroupMembershipService`.  This prevents a circular dependency forming, but more importantly it stops a possible overreach of the `ReferralService` to becoming a "God Object" which does everything.  By distributing functionality into the domain of `ProgrammeGroupMembership`, there remains a one-way dependency (Referral -> ProgrammeGroupMembership) but the dependencies of the `ProgrammeGroupMembershipService` remain _purely_ a set of Repositories, which _feels_ cleaner.

* APG-1507: Create the endpoint to remove a Referral from a Group

**WHAT**

This commit introduces the `POST /group/:id/remove/:referral_id`, which
links to the previously introduced
`ProgrammeGroupMembershipService::removeFromGroup` method

**WHY**

Before this commit, the Service method was only used in test.  This
commit plugs the method into the public-facing API.

* APG-1507: Make the additional_details on the Remove From Group endpoint optional

* APG-1507: fix duplicates from rebase