APG-1425: Filter (Programme)Groups by PDU & Delivery Location (#332)
* APG-1425: Filter (Programme)Groups by PDU & Delivery Location

**WHAT**

Most simply this commit removes a dangling `/region` part of a URL,
which does not function for anything.

This commit makes changes to the /bff/programme-groups endpoint, by
returning the list of both PDUs and Reporting Locations of the Groups.

It also uses a single PDU and possibly a set of Delivery Locations (both
nullable) to filter the ProgrammeGroups.

**WHY**

The BFF endpoint needs to support the Group List page, which needs to be
presented with a complete list of PDUs, as well as a list of Delivery
Locations of the Referrals.

This requires the endpoint to return a list of both PDUs and Delivery
Locations, which the front-end can take to feed into the relevant
`select` components and checkboxes.

**REFACTORING**

The commit also changes the PDU of the Group in the Local migration to
be in the same Region as the wiremock will return.  Without this, the
Group list is empty.

# Conflicts:
#	src/main/kotlin/uk/gov/justice/digital/hmpps/accreditedprogrammesmanageanddeliverapi/service/ProgrammeGroupService.kt

* APG-1425: Update Integration tests on the ProgrammeGroupService

* APG-1425: shift references to the EntityManager out of the ProgrammeGrouService and towards a Custom Repo

* APG-1425: Remove custom ProgrammeGroupRepository implementation; replace with @Query methods to simplify distinct field value retrieval

---------

Co-authored-by: pfcairns <phil.feroldi-cairns@digital.justice.gov.uk>