endpoint: POST /group/:id/session/schedule
**WHAT**

This commit introduces a new POST endpoint at the above URL.

It includes Swagger documentation and request validation.

It does not implement the logic of scheduling the session.

**WHY**

This is an incremental delivery commit, where the boilerplate for
building the endpoint, as well as its validation, is included.

This will allow our UI to plug into _something_ as we build the
multi-page form for scheduling a specific one-to-one meeting for a
Member of a Group

**HOW**

This commit follows codebase and Spring Boot practices by placing the
endpoint in the ProgrammeGroupController, and some tests in the
Integration Tests for that controller.

It leaves a comment explaining where the business logic will be
referenced.