Endpoint: /bff/group/:group_id/module/:module_id/schedule-individual-session-details
**WHAT**

This commit introduces a new endpoint at:

`GET /bff/group/:group_id/module/:module_id/schedule-individual-session-details`

Which return a list of Group Members and Facilitators for the specified
Group.

**WHY**

This BFF feeds data into screen two of a multi-page form for scheduling
Sessions for Members of a Group.  At present, this is purely one-to-one Sessions.

In the future, we will be using this to schedule one-to-one and Group catch-up
Sessions.  This is why we need the `:module_id`, it's a small bit of
information that allows us to future-proof.

**HOW**

This commit uses largely existing functionality.  With the exception of
a new ProgrammeGroupMembershipRepository method which fetches all of the
current active Members of a Group.

The commit introduces the endpoint and its documentation, and provides
two simple tests (one happy, one unhappy path) for testing it.