Create BFF for loading One-to-One sessions for a Group/Module (#347)
* endpoint: /bff/group/{groupId}/module/{moduleId}/schedule-session-type

**WHAT**

This commit introduces the following endpoint:

`POST /bff/group/{groupId}/module/{moduleId}/schedule-session-type`

Which returns a list of (One-to-One) Session Templates for the specified
Group/Module combination.

In doing so, the `TemplateService` was created which does the fetching
and filtering of entities.

**WHY**

This endpoint will be used as on the first of a multi-page form which is
used for scheduling one-to-one sessions for Members of a Group.  To do
that, the UI needs to know about the possible One-to-One sessions which
could exist.

**HOW**

- New endpoint is created and documented with Swagger docs
- Created the `TemplateService` to find the entities in the database.
- Added Integration tests to the Controller layer, and more to the
  Service layer.

# Conflicts:
#	src/main/kotlin/uk/gov/justice/digital/hmpps/accreditedprogrammesmanageanddeliverapi/api/controller/ProgrammeGroupController.kt
#	src/test/kotlin/uk/gov/justice/digital/hmpps/accreditedprogrammesmanageanddeliverapi/api/controller/ProgrammeGroupControllerIntegrationTest.kt

* docs: Add an explanatory comment to the /bff/group/{groupId}/module/{moduleId}/schedule-session-type endpoint

* fix: Changes ProgrammeGroupController response type in tests; Change return type of method in GroupService